backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
faster-whisper==1.0.1
openai==1.12.0
# pyannote.audio==3.1.1  # Disabled - requires cmake (add later for diarization)
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pydub==0.25.1
